chore: update character models
No idea what is actually different here. The thing is, there is also a character model for My First Game, Barky. I also tried saving the original BloxCode Killer from the real game and it shows up as being a different file from the one in this commit, but it still shows up as a change, meaning it is different from the file in the remote repository before this commit. Maybe I syncbacked weirdly?