style: remove `deviation` script
I would have kept it, but it kept getting marked as a change whenever I
pulled changes from Studio, and then I realised it is useless anyway
since nobody cares if those models were folders + there are other
deviations that I did not record, so I am deciding now to remove this
file.

Although this has nothing to do with style, it does fit the definition
of a commit that does not change behaviour at all, since this code file
entirely consisted of a comment, so I think it fits the `style` type of
Conventional Commits.