build: add Lune script for syncback
Not sure if I will use it since it automatically confirms the syncback
operation. It would be nice to be able to see how many files will be
changed during the operation before confirming. Thus, I am not sure if
I will use this script, but for now I will try it.