build: add Rojo to `rokit.toml`
Add both `rojo` and `rojo-syncback`. I think `rojo-syncback` can be used
for regular Rojo stuff as well, but just in case I want to use it only
for syncback.

Rojo was already being used for this project, as seen in the Lune
scripts, so this commit is kind of late, but Better Late Than Never.