build: add separate syncback file for archiving
It only archives the `ServerStorage/OldGames` folder. It may be
beneficial to add the option to specify which game in particular to
archive. Unfortunately, it appears as though there is no way to specify
to only archive a certain folder, so instead the file says to ignore
every other path in ServerStorage.